AI-Dungeon-Master-main/requirements.txt
# Core AI Libraries
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.0.350
tiktoken>=0.5.0

# Web Framework
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Game Logic
dice>=4.0.0
random-word>=1.0.11

# Utility Libraries
requests>=2.31.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced AI features
transformers>=4.35.0
torch>=2.0.0

# Optional: For voice features
speechrecognition>=3.10.0
gtts>=2.4.0

# Optional: For image generation
pillow>=10.0.0